pyproject.toml
[project]
name = "a2a-101"
version = "0.1.0"
description = "Agent-to-Agent Communication Protocol starter project"
keywords = ["a2a", "agent", "communication", "protocol"]
authors = [
    {name = "Ndamulelo Nemakhavhani",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "a2a-sdk (>=0.3.0,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
